requirements.txt
# Core
PyYAML>=6.0

# Testing
pytest>=8.0
pytest-cov>=5.0

# Optional but recommended for type safety
typing-extensions>=4.8